app/backend/requirements.txt
azure-identity==1.13.0b3
Flask==2.2.2
langchain==0.0.78
openai==0.26.4
azure-storage-blob==12.14.1
pinecone-client==2.2.1
sentence-transformers==2.2.2
torch==2.0.1
